add sorting to all charts
